app/backend/db/requirements.txt
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0